change structure of lib + specs
