+++ spec/spec_helper.rb
@@ -1,4 +1,5 @@
-require_relative "../lib/ruby_collaborating_objects_lab.rb"
+require_relative "../lib/artist.rb"
+require_relative "../lib/song.rb"
 
 # This file was generated by the `rspec --init` command. Conventionally, all
 # specs live under a `spec` directory, which RSpec adds to the `$LOAD_PATH`.
